requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
anthropic==0.44.0
streamlit==1.41.1
python-multipart==0.0.20
fpdf2==2.8.2
python-dotenv==1.0.1
httpx==0.28.1
